rename ingest/metadata, rearrange slightly
